perf(cli): defer health text rendering until success
